tools/requirements.txt
# Jinja2 Static Site Generator Dependencies
# Install with: pip install -r requirements.txt

Jinja2==3.1.3
PyYAML==6.0.1
MarkupSafe==2.1.5
